requirements.txt
fastapi
sqlalchemy
python-multipart
requests
uvicorn
passlib[bcrypt]
python-jose[cryptography]
google-auth
twilio
python-dotenv
